PDF_Studio/requirements.txt.txt
flask
flask-sqlalchemy
pdfplumber
pypdf
python-docx
gunicorn